Temporarily revert prevent_destroy.
Combining `prevent_destroy` with `ignore_changes` on a dynamic variable
effectively makes it impossible to update the relevant resource, as
described in https://github.com/hashicorp/terraform/issues/7728. This
patch reverts the `prevent_destroy` flag on the rds resource. If the
underlying issue is fixed, we can restore this flag. Otherwise, we'll
have to either specify all database identifiers or allow them to be
randomly generated.