use the VPC flow log Terraform module
This reusable module was built as part of the DevSecOps effort, and we
are looking for adoption and feedback. Rather than every project needing
to include the same boilerplate code (for setting up the flow logs, in
this case), hoping to get various projects at GSA to use reusable pieces
like this one. Let me know what you think!

I `valididate`d the module locally, but don't have the ability to test
it for real. Since the `aws_cloudwatch_log_group` changed modules, you
will need to `terraform state mv` it to ensure the log group isn't
destroyed and recreated.