Fix the path issues for $out_path;
Seems like `$out_path` now includes a `README` file, so `cp` is having
issues finding the files. I've tested this in a Concourse worker to
determine that we should only be listing directories from the `live/`
directory and skip over any of the files in there such as `README`.

_This is brittle enough that it will break again if there's another
directory._