Update Terraform-managed RDS instances - bosh (spoke)
This changeset is part of a multi-step series of actions to upgrade our Terraform-managed RDS instances that are internal to the cloud.gov platform.  When all steps are completed we should see the following:

- All RDS instances running on the latest version available
- RDS instances down-sized to db.t3.micro instances where appropriate, or db.t3.small/db.t3.medium where still necessary

In order to accomplish this, the instances have to be upgraded in phases since you cannot go from a 9.6.x release straight to a 12.x release.  Furthermore, the db.t3.* instance class is not available for anything below PostgreSQL 11.5 (all internal databases are PostgreSQL instances).  In our other work updating and down-sizing other PostgreSQL instances, we have found that we can upgrade from a 9.6.x release to 11.1, then go to the latest 12.x release (12.3 at the time this PR was opened), provided there are no extra extensions enabled or undefined field types.