Address issues with broker failing when user is missing
The s3 broker fails to unbind if the underlying IAM user has been deleted or somehow failed to be instantiated.

To prevent AccessDenied issues, we'll let ListAccessKeys and ListAttachedUsersPolicies fail with a NotFound by granting them access to all users.

The alternative would be more permissions for GetUsers or ListUsers, which would require a code rewrite, and those permissions  leak more information if this IAM role were to be compromised